fix bug with move counter when undoing
clicking on the current colour would add history entries that would decrement the total moves when undo was clicked. Consolidated change colour and increment moves state actions into one and added check in reducer instead of in the functional stateless component to see if the clicked colour was the current player colour.